moved static impl to Monitors as a static method
