+++ src/test/java/com/flipkart/ranger/healthservice/monitor/ServiceProviderInlineMonitorTest.java
@@ -140,7 +140,7 @@ public class ServiceProviderInlineMonitorTest {
                         return HealthcheckStatus.healthy;
                     }
                 })
-                .withInlineHealthMonitor(new Monitor.FileExistanceCheckMonitor(file.getAbsolutePath()))
+                .withInlineHealthMonitor(Monitors.fileExistanceCheckMonitor(file.getAbsolutePath()))
                 .buildServiceDiscovery();
         serviceProvider.start();
     }
